backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
idna==3.4
kombu==5.3.2
mysqlclient==2.2.0
numpy==1.26.1
pandas==2.1.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
python-decouple
psycopg2-binary
django-celery-beat